refactor: move blueprint codes into one
