+++ simaple/gear/spec/builtin/Legendary.yaml
