+++ simaple/gear/spec/builtin/Legendary18.yaml
