+++ simaple/gear/spec/builtin/LegendaryHalf.yaml
