+++ simaple/gear/spec/resources/item_name_alias.json
